requirements.txt
# FastAPI web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database ORM
sqlalchemy==2.0.25

# Templates and forms
jinja2==3.1.3
python-multipart==0.0.6

# HTTP requests (for Telegram)
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Input device handling (RFID reader)
evdev==1.6.1

# GPIO control for Raspberry Pi
RPi.GPIO==0.7.1
